refactor(#174): reorganizes folder structure and types
