+++ packages/react-dadata/setupTests.ts
@@ -1,3 +1,4 @@
+/// <reference types="@testing-library/jest-dom/vitest" />
 import '@testing-library/jest-dom/vitest';
 import { cleanup } from '@testing-library/react';
 import { afterEach } from 'vitest';
