+++ packages/react-dadata/src/index.tsx
@@ -1,79 +1,65 @@
-import { AddressSuggestions } from './AddressSuggestions';
-import { BankSuggestions } from './BankSuggestions';
-import { EmailSuggestions } from './EmailSuggestions';
-import { FioSuggestions } from './FioSuggestions';
-import { PartySuggestions } from './PartySuggestions';
-import type {
-  DaDataAddress,
-  DaDataAddressBounds,
-  DaDataBank,
-  DaDataBankStatus,
-  DaDataBankType,
-  DaDataEmail,
-  DaDataFio,
-  DaDataGender,
-  DaDataParty,
-  DaDataPartyBelarus,
-  DaDataPartyBelarusStatus,
-  DaDataPartyBranchType,
-  DaDataPartyRussia,
-  DaDataPartyRussiaStatus,
-  DaDataPartyStatus,
-  DaDataPartyType,
-  DaDataSuggestion,
-} from './types';
-import { PartyBelarusSuggestions } from './variants/party_belarus/party-belarus';
-import { PartyKazakhstanSuggestions } from './variants/party_kazakhstan/party-kazakhstan';
-import type {
-  DaDataPartyKazakhstan,
-  DaDataPartyKazakhstanStatus,
-  DaDataPartyKazakhstanType,
-} from './variants/party_kazakhstan/party-kazakhstan-types';
-
-type DaDataAddressSuggestion = DaDataSuggestion<DaDataAddress>;
-type DaDataPartySuggestion = DaDataSuggestion<DaDataParty>;
-type DaDataPartyRussiaSuggestion = DaDataSuggestion<DaDataPartyRussia>;
-type DaDataPartyBelarusSuggestion = DaDataSuggestion<DaDataPartyBelarus>;
-type DaDataPartyKazakhstanSuggestion = DaDataSuggestion<DaDataPartyKazakhstan>;
-type DaDataBankSuggestion = DaDataSuggestion<DaDataBank>;
-type DaDataFioSuggestion = DaDataSuggestion<DaDataFio>;
-type DaDataEmailSuggestion = DaDataSuggestion<DaDataEmail>;
-
-export {
-  type DaDataSuggestion,
-  AddressSuggestions,
-  PartySuggestions,
-  PartyBelarusSuggestions,
-  PartyKazakhstanSuggestions,
-  BankSuggestions,
-  FioSuggestions,
-  EmailSuggestions,
-  type DaDataAddress,
-  type DaDataAddressSuggestion,
-  type DaDataParty,
-  type DaDataPartySuggestion,
-  type DaDataPartyRussiaSuggestion,
-  type DaDataPartyBelarusSuggestion,
-  type DaDataPartyKazakhstanSuggestion,
-  type DaDataAddressBounds,
-  type DaDataPartyType,
-  type DaDataPartyRussia,
-  type DaDataPartyRussiaStatus,
-  type DaDataPartyBranchType,
-  type DaDataPartyStatus,
-  type DaDataPartyBelarus,
-  type DaDataPartyBelarusStatus,
-  type DaDataPartyKazakhstan,
-  type DaDataPartyKazakhstanStatus,
-  type DaDataPartyKazakhstanType,
-  type DaDataBank,
-  type DaDataBankStatus,
-  type DaDataBankType,
-  type DaDataBankSuggestion,
-  type DaDataFio,
-  type DaDataFioSuggestion,
-  type DaDataGender,
-  type DaDataEmail,
-  type DaDataEmailSuggestion,
-};
+/**
+ * Общие типы
+ */
+export type { DaDataSuggestion } from './core-types';
 export { HttpCache } from './http-cache';
+
+/**
+ * Адреса
+ */
+export type { DaDataAddress, DaDataAddressBounds, DaDataAddressSuggestion } from './variants/address/address-types';
+export { AddressSuggestions } from './variants/address/address-suggestions';
+
+/**
+ * Организации в России 🇷🇺
+ */
+export type {
+  DaDataPartyRussia,
+  DaDataPartyType,
+  DaDataPartyStatus,
+  DaDataPartyBranchType,
+  DaDataPartyRussiaFio,
+  DaDataPartyRussiaSuggestion,
+  DaDataPartySuggestion,
+} from './variants/party_russia/party-russia-types';
+export { PartySuggestions } from './variants/party_russia/party-russia-suggestions';
+
+/**
+ * Организации в Беларуси 🇧🇾
+ */
+export type {
+  DaDataPartyBelarus,
+  DaDataPartyBelarusType,
+  DaDataPartyBelarusStatus,
+  DaDataPartyBelarusSuggestion,
+} from './variants/party_belarus/party-belarus-types';
+export { PartyBelarusSuggestions } from './variants/party_belarus/party-belarus-suggestions';
+
+/**
+ * Организации в Казахстане 🇰🇿
+ */
+export type {
+  DaDataPartyKazakhstan,
+  DaDataPartyKazakhstanSuggestion,
+  DaDataPartyKazakhstanType,
+  DaDataPartyKazakhstanStatus,
+} from './variants/party_kazakhstan/party-kazakhstan-types';
+export { PartyKazakhstanSuggestions } from './variants/party_kazakhstan/party-kazakhstan-suggestions';
+
+/**
+ * Банки
+ */
+export type { DaDataBank, DaDataBankStatus, DaDataBankType, DaDataBankSuggestion } from './variants/bank/bank-types';
+export { BankSuggestions } from './variants/bank/bank-suggestions';
+
+/**
+ * ФИО
+ */
+export type { DaDataFio, DaDataFioSuggestion } from './variants/fio/fio-types';
+export { FioSuggestions } from './variants/fio/fio-suggestions';
+
+/**
+ * Email
+ */
+export type { DaDataEmail, DaDataEmailSuggestion } from './variants/email/email-types';
+export { EmailSuggestions } from './variants/email/email-suggestions';
